requirements.txt
requests==2.26.0
Pillow==8.3.1
openai==0.27.0
colorama==0.4.4